rename core --> Cosmology
