+++ pyccl/base/repr_.py
@@ -114,7 +114,7 @@ def build_string_Cosmology(self):
 
     Example output ::
 
-        <pyccl.core.Cosmology>
+        <pyccl.cosmology.Cosmology>
             Omega_b = 0.05
             Omega_c = 0.25
             h       = 0.67
@@ -178,7 +178,7 @@ def build_string_Cosmology(self):
             s += f"{newline}HASH_PK = {H}"
         return s
 
-    s = "<pyccl.core.Cosmology>"
+    s = "<pyccl.cosmology.Cosmology>"
     s += printdict(self._params_init_kwargs)
     s += printdict(self._config_init_kwargs)
     s += printextras(self._params_init_kwargs)
